The __wait_for_all function was fixed
Now the system is launched into the SM task without raising an error due to missing data. The system feels slightly disorganized code-wise, but will execute everything in the order that will be final, and if we measure the runtime and periodicity these will be accurate to the final product due to the CAN loopback.
Maybe some work could be done to make it more obvious how tasks interact/share data